moving runRound to round class
